% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cases_table.R
\name{create_cases_table}
\alias{create_cases_table}
\title{Create the daily or monthly COVID-19 cases table of a country or a province}
\usage{
create_cases_table(country_select, province_select = "", cases_frequency)
}
\arguments{
\item{country_select}{The name of a country as a character vector.
Currently, there are three countries available, which are "Japan", "US", and "Australia".}

\item{province_select}{The name of a province as a character vector.
Currently, only Australia has the provincial data. The available provinces are
"Australian Capital Territory", "New South Wales", "Northern Territory", "Queensland",
"South Australia", "Tasmania", "Victoria", and "Western Australia".}

\item{cases_frequency}{A character vector indicates the frequency of cases in the table.
The available choices are "monthly" or "daily".}
}
\value{
A table that shows the daily or monthly COVID-19 cases and the cumulative cases of a country or a province.
}
\description{
This function is used to create the table to display the daily
or monthly COVID-19 cases of a inputted country or province. At the same time,
the cumulative number will also be displayed in each table.
This function refactors one part of the server-side logic of the Shiny app,
which is the table of displaying the daily or monthly COVID-19 cases of a country or a province.
}
\note{
Only Australia has the available provincial data,
please do not input any character string into the argument of \emph{province_select} when choosing "Japan" or "US".
}
\examples{
\dontrun{

# Create the monthly table of Japan.

create_cases_table(country_select = "Japan", cases_frequency = "monthly")

# Create the monthly table of Victoria, Australia.

create_cases_table(country_select = "Australia", 
                   province_select = "Victoria", 
                   cases_frequency = "monthly")
                   
# Create the daily table of US.                  
                   
create_cases_table(country_select = "US", cases_frequency = "daily")

}

}
