% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny app}
\source{
The Shiny app could be found in \href{https://jinhao-luo.shinyapps.io/shiny-assessment-jluo0015/}{here}
}
\usage{
launch_app()
}
\value{
The Shiny app will run automatically, and the interface of the app will pop up.
}
\description{
This function is used to launch the Shiny app of the Shiny assessment,
where is placed in the inst/app folder.
}
\note{
This function does not require any arguments to run this function. Simply enter and run this function.
}
\examples{
\dontrun{

# Use the following function to run the app

launch_app()

}

}
