% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchshiny-package.R
\docType{package}
\name{launchshiny-package}
\alias{launchshiny}
\alias{launchshiny-package}
\title{launchshiny: Launch the Shiny App by a Function of the Launchshiny Package}
\description{
This package is used to launch the Shiny app of the Shiny assessment by a function. Meanwhile, some logics of app would be refactored into R functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-jluo0015}
  \item Report bugs at \url{https://github.com/etc5523-2020/r-package-assessment-jluo0015/issues}
}

}
\author{
\strong{Maintainer}: Jinhao Luo \email{jluo0015@student.monash.edu} (\href{https://orcid.org/0000-0001-6702-1023}{ORCID})

}
\keyword{internal}
