% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_daily_cases.R
\name{plot_country_daily_cases}
\alias{plot_country_daily_cases}
\title{Create the line chart of daily COVID-19 cases of a country in a month}
\usage{
plot_country_daily_cases(country_select, month_select)
}
\arguments{
\item{country_select}{The name of a country as a character vector.
Currently, there are three countries available, which are "Japan", "US", and "Australia".}

\item{month_select}{The number of a month as an integer vector.
The available number of month is between 1 and 10.}
}
\value{
A line graph that shows three types of daily COVID-19 cases for a month in a country.
}
\description{
This function is used to create a line chart to display the daily COVID-19
situation of a inputted country in a inputted month.
This line chart will show three types of COVID-19 cases. And it is a plotly line chart.
This function refactors one part of the server-side logic of the Shiny app,
which is the plot of displaying the daily COVID-19 cases of a country in a specific month.
}
\examples{
\dontrun{
plot_country_daily_cases("Japan", 7)

plot_country_daily_cases("Australia", 4)

plot_country_daily_cases("US", 10)
}

}
