% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_country_month.R
\name{select_country_month}
\alias{select_country_month}
\title{Select the daily COVID-19 cases of a country within a month}
\format{
This function will return a tibble with 7 columns.

\strong{country}
\itemize{
\item Name of the country.
}

\strong{year}, \strong{month}, \strong{day}
\itemize{
\item The number of the year, month and day.
}

\strong{confirmed}
\itemize{
\item The number of daily confirmed cases of COVID-19.
}

\strong{death}
\itemize{
\item The number of daily mortality of COVID-19.
}

\strong{recovered}
\itemize{
\item The number of daily recovered cases of COVID-19.
}
}
\usage{
select_country_month(country_select, month_select)
}
\arguments{
\item{country_select}{The name of a country as a character vector.
Currently, there are three countries available, which are "Japan", "US", and "Australia".}

\item{month_select}{The number of a month as an integer vector.
The available number of month is between 1 and 10.}
}
\value{
A tibble that contains the information of country, year, month, day, and daily COVID-19 cases.
}
\description{
This function is used to find the three different daily
COVID-19 cases of the selected country in a specific month.
This function refactors one part of the user interface-side (UI) logic of the Shiny app,
which is the selection bar of selecting the country and a specific month.
}
\examples{
select_country_month(country_select = "Japan", month_select = 1)

select_country_month(country_select = "US", month_select = 3)

select_country_month(country_select = "Australia", month_select = 5)

}
