% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny app}
\usage{
launch_app()
}
\arguments{
\item{NULL}{do not require any arguments to run this function. Simply enter and run this function.}
}
\value{
The Shiny app will run automatically, and the interface of the app will pop up.
}
\description{
This function is used to launch the Shiny app of the Shiny assessment,
where is placed in the inst/app folder.
}
\examples{
launch_app()

}
